% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_HRbound.R
\name{boot_HRbound}
\alias{boot_HRbound}
\title{Bootstrap Harm Rate Bound}
\usage{
boot_HRbound(Tr, S, Y, c1 = 0, c3 = 1, nboot, conf.level = 0.95)
}
\arguments{
\item{Tr}{Treatment variable, a vector taking value 1 or 0.}

\item{S}{Surrogate variable, a vector taking value 1 or 0.}

\item{Y}{Outcome variable, a vector taking value 1 or 0.}

\item{c1}{0<=c1<=1, representing the harm rate from the treatment to the surrogate.}

\item{c3}{0<=c3<=1, representing the upper bound for the propotion of the individuals that do not conform to causal necessity.}

\item{nboot}{Number of bootstrap samples. A positive integer.}

\item{conf.level}{Confidence level of the interval.}
}
\value{
Calculate confidence interval for the bound of the hard rate from the treatment to the outcome.
}
\description{
Calculate bootstrap confidence interval for the bound of harm rate from the treatment to the outcome.
}
\details{
This function computes the bootstrap confidence interval for the sharp bound of harm rate from the treatment to the surrogate.
}
\examples{
boot_HRbound(Tr = ACCORD_eye[, 1], S = ACCORD_eye[, 2], Y = ACCORD_eye[, 3], n = 1000)
}
