% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRbound.R
\name{HRbound}
\alias{HRbound}
\title{Harm Rate Bounds}
\usage{
HRbound(Tr, S, Y, c1 = 0, c3 = 1)
}
\arguments{
\item{Tr}{Treatment variable, a vector taking value 1 or 0.}

\item{S}{Surrogate variable, a vector taking value 1 or 0.}

\item{Y}{Outcome variable, a vector taking value 1 or 0.}

\item{c1}{0<=c1<=1, representing the harm rate from the treatment to the surrogate.}

\item{c3}{0<=c3<=1, representing the upper bound for the propotion of the individuals that do not conform to causal necessity.}
}
\value{
Lower and upper bound of the harm rate from the treatment to the outcome.
}
\description{
Calculate the bounds of harm rate from the treatment to the outcome.
}
\details{
This function calculates the lower and upper sharp bounds of harm rate from the treatment to the surrogate, using the method proposed in Ma et al.
}
\examples{
HRbound(Tr = ACCORD_eye[, 1], S = ACCORD_eye[, 2], Y = ACCORD_eye[, 3])
}
\references{
Ma, L., Yin, Y., Liu, L. and Geng, Z. On the Individual Surrogate Paradox. Submitted
}
