% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISP.R
\name{ISP}
\alias{ISP}
\title{Individual Surrogate Paradox}
\usage{
ISP(Tr, S, Y, c1, c3 = 1)
}
\arguments{
\item{Tr}{Treatment variable, a vector taking value 1 or 0.}

\item{S}{Surrogate variable, a vector taking value 1 or 0.}

\item{Y}{Outcome variable, a vector taking value 1 or 0.}

\item{c1}{0<=c1<=1, representing the harm rate from the treatment to the surrogate.}

\item{c3}{0<=c3<=1, representing the upper bound for the propotion of the individuals that do not conform to causal necessity.}
}
\value{
Lower and upper bound of the harm rate from the treatment to the outcome, and whether the individual surrogate paradox manifests.
}
\description{
Check whether individual surrogate paradox manifests by calculating the bound of harm rate from the treatment to the outcome.
}
\details{
This function calculates the sharp bounds of harm rate from the treatment to the surrogate, and determine whether individual surrogate paradox exists using the method proposed in Ma et al.
}
\examples{
ISP(Tr = ACCORD_eye[, 1], S = ACCORD_eye[, 2], Y = ACCORD_eye[, 3], c1 = 0.5)
}
\references{
Ma, L., Yin, Y., Liu, L. and Geng, Z. On the Individual Surrogate Paradox. Submitted
}
