% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bound.R
\name{plot_bound}
\alias{plot_bound}
\title{Plot Harm Rate Bounds}
\usage{
plot_bound(Tr, S, Y, c3 = 1)
}
\arguments{
\item{Tr}{Treatment variable, a vector taking value 1 or 0.}

\item{S}{Surrogate variable, a vector taking value 1 or 0.}

\item{Y}{Outcome variable, a vector taking value 1 or 0.}

\item{c3}{0<=c3<=1, representing the upper bound for the propotion of the individuals that do not conform to causal necessity.}
}
\value{
Lower and upper bound of the harm rate from the treatment to the outcome, and whether the individual surrogate paradox manifests.
}
\description{
Present a plot for the bounds of harm rate from the treatment to the outcome.
}
\details{
This function plots the lower and upper sharp bounds of harm rate from the treatment to the surrogate when c1 varies.
}
\examples{
plot_bound(Tr = ACCORD_eye[, 1], S = ACCORD_eye[, 2], Y = ACCORD_eye[, 3], c3 = 0)
}
\references{
Ma, L., Yin, Y., Liu, L. and Geng, Z. On the Individual Surrogate Paradox. Submitted
}
